Fix compilation warnings and test suite macro redefinitions
* Resolves signedness comparison warnings in `kernel/fs/devfs.c` (BINDER ioctls).
* Resolves unused parameter warnings in `sys_umask` in `kernel/arch/x86_64/syscall.c`.
* Silences unused static wrapper functions in `kernel/arch/x86_64/syscall.c`.
* Resolves macro redefinitions (`__ETEROS_HOST_TEST__`, `PROT_READ`, `HAL_MEM_WRITE`, etc.) in `test_stack_security.c`, `test_mmap_fixed.c`, and `test_framebuffer_scroll.c`.
* Removes `|| true` from test scripts to enforce CI validation.
* Updates Orchestrator Report and assigned agents (`aether-droid-subsystem-bot`, `network-socket-api-bot`, `testing-ci-validation-bot`, `userspace-libc-posix-bot`) to "Waiting for new assignment".

+++ kernel/fs/devfs.c
@@ -87,9 +87,9 @@ static uint32_t dev_null_write(fs_node_t *node, uint32_t offset, uint32_t size, 
 /* /dev/binder Implementation (Stub for Android Compatibility)               */
 /* ========================================================================= */
 
-#define BINDER_VERSION_IOWR 0xc0046209 /* Linux ioctl code for BINDER_VERSION */
-#define BINDER_WRITE_READ 0xc0306201
-#define BINDER_SET_CONTEXT_MGR 0x40046207
+#define BINDER_VERSION_IOWR (int)0xc0046209 /* Linux ioctl code for BINDER_VERSION */
+#define BINDER_WRITE_READ (int)0xc0306201
+#define BINDER_SET_CONTEXT_MGR (int)0x40046207
 
 struct binder_version {
     int32_t protocol_version;
